+++ fluidpythran/aheadoftime.py
@@ -27,6 +27,8 @@ import inspect
 import time
 import subprocess
 import os
+import functools
+import sys
 
 from .util import (
     get_module_name,
@@ -36,7 +38,11 @@ from .util import (
     modification_date,
 )
 
-from .pythranizer import compile_pythran_file, name_ext_from_path_pythran
+from .pythranizer import (
+    compile_pythran_file,
+    name_ext_from_path_pythran,
+    ext_suffix,
+)
 
 from .annotation import (
     make_signature_from_template_variables,
@@ -79,7 +85,7 @@ def _get_fluidpythran_calling_module(index_frame: int = 2):
     if module_name in modules:
         fp = modules[module_name]
         if (
-            fp._created_while_transpiling != is_transpiling
+            fp.is_transpiling != is_transpiling
             or fp._pythranize_at_import_at_creation
             != has_to_pythranize_at_import()
             or hasattr(fp, "path_mod")
@@ -196,7 +202,7 @@ class FluidPythran:
 
         self.names_template_variables = {}
 
-        self._created_while_transpiling = is_transpiling
+        self.is_transpiling = is_transpiling
 
         if is_transpiling:
             self.functions = {}
@@ -218,11 +224,11 @@ class FluidPythran:
             module_short_name = module_name
             module_pythran_name = ""
 
-        module_pythran_name += "__pythran__._" + module_short_name
+        module_pythran_name += "__pythran__." + module_short_name
 
         self.path_mod = path_mod = Path(frame.filename)
         self.path_pythran = path_pythran = (
-            path_mod.parent / "__pythran__" / ("_" + module_short_name + ".py")
+            path_mod.parent / "__pythran__" / (module_short_name + ".py")
         )
 
         path_ext = None
@@ -240,7 +246,13 @@ class FluidPythran:
                     # better to do this in another process because the file is already run...
                     os.environ["FLUIDPYTHRAN_NO_MPI"] = "1"
                     returncode = subprocess.call(
-                        ["fluidpythran", "-np", str(path_mod)]
+                        [
+                            sys.executable,
+                            "-m",
+                            "fluidpythran.run",
+                            "-np",
+                            str(path_mod),
+                        ]
                     )
                     del os.environ["FLUIDPYTHRAN_NO_MPI"]
                 returncode = mpi.bcast(returncode)
@@ -288,6 +300,12 @@ class FluidPythran:
             self.is_compiled = False
 
         self.is_transpiled = True
+
+        if not path_ext.exists() and not self.is_compiling:
+            path_ext_alt = path_pythran.with_suffix(ext_suffix)
+            if path_ext_alt.exists():
+                self.path_extension = path_ext = path_ext_alt
+
         if path_ext.exists() and not self.is_compiling:
             self.module_pythran = import_from_path(
                 self.path_extension, module_pythran_name
@@ -338,7 +356,7 @@ class FluidPythran:
             func_tmp = func
 
         if self.is_compiling:
-            return CheckCompiling(self, func_tmp)
+            return functools.wraps(func)(CheckCompiling(self, func_tmp))
 
         return func_tmp