+++ fluidpythran/justintime.py
@@ -59,6 +59,7 @@ import os
 import sys
 import time
 from distutils.util import strtobool
+from functools import wraps
 
 try:
     import numpy as np
@@ -88,7 +89,9 @@ modules = {}
 
 
 path_cachedjit = mpi.Path(path_root) / "__cachedjit__"
-path_cachedjit.mkdir(parents=True, exist_ok=True)
+if mpi.rank == 0:
+    path_cachedjit.mkdir(parents=True, exist_ok=True)
+mpi.barrier()
 
 _COMPILE_CACHEDJIT = strtobool(os.environ.get("FLUID_COMPILE_CACHEDJIT", "True"))
 
@@ -252,7 +255,10 @@ class CachedJIT:
         module_name = mod.module_name
 
         path_pythran = path_cachedjit / module_name.replace(".", os.path.sep)
-        path_pythran.mkdir(parents=True, exist_ok=True)
+
+        if mpi.rank == 0:
+            path_pythran.mkdir(parents=True, exist_ok=True)
+        mpi.barrier()
 
         path_pythran = (path_pythran / func_name).with_suffix(".py")
         path_pythran_header = path_pythran.with_suffix(".pythran")
@@ -389,6 +395,7 @@ class CachedJIT:
             module_pythran = import_from_path(path_ext, name_mod)
             self.pythran_func = getattr(module_pythran, func_name)
 
+        @wraps(func)
         def type_collector(*args, **kwargs):
 
             if self.compiling: