+++ fluidpythran/pythranizer.py
@@ -29,6 +29,7 @@ import time
 from typing import Union, Iterable, Optional
 import sysconfig
 import hashlib
+import sys
 
 from .compat import open, implementation
 from . import mpi
@@ -135,7 +136,12 @@ class SchedulerPopen:
         if openmp:
             update_flags("-fopenmp")
 
-        words_command = ["_pythran-fluid", path.name]
+        words_command = [
+            sys.executable,
+            "-m",
+            "fluidpythran_cl.run_pythran",
+            path.name,
+        ]
 
         if name_ext_file is None:
             name_ext_file = name_ext_from_path_pythran(path)
@@ -175,7 +181,7 @@ def compile_pythran_files(
     paths: Iterable[Path], str_pythran_flags: str, parallel=True
 ):
     for path in paths:
-        print("pythranize file", path)
+        print("Schedule pythranization of file", path)
         scheduler.compile_with_pythran(
             path, str_pythran_flags=str_pythran_flags, parallel=parallel
         )