+++ fluidpythran/run.py
@@ -8,15 +8,11 @@ Internal API
 
 .. autofunction:: parse_args
 
-.. autofunction:: pythran_fluid
-
 """
 
 import argparse
 from pathlib import Path
 from glob import glob
-import subprocess
-import sys
 
 from . import __version__
 from .transpiler import make_pythran_files
@@ -87,7 +83,7 @@ def run():
     pythran_paths = []
     for path in paths:
         path = Path(path)
-        pythran_path = path.parent / "__pythran__" / ("_" + path.name)
+        pythran_path = path.parent / "__pythran__" / path.name
         ext_path = pythran_path.with_suffix(ext_suffix)
         if has_to_build(ext_path, pythran_path):
             pythran_paths.append(pythran_path)
@@ -144,3 +140,7 @@ def parse_args():
     )
 
     return parser.parse_args()
+
+
+if __name__ == "__main__":
+    run()