+++ fluidpythran/test_init.py
@@ -1,4 +1,3 @@
-
 import importlib
 
 from . import FluidPythran, mpi